requirements.txt
Django==1.8.5
django-bower==5.0.4
djangorestframework==3.2.4
drf-nested-routers==0.10.0
six==1.10.0
wheel==0.24.0
